requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
idna==3.3
oauthlib==3.2.0
pkg_resources==0.0.0
python-dotenv==0.20.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
urllib3==1.26.11
